Revert "chore: add action updater"
This reverts commit 133e832a8c9d5770656fe856571f2a7c941108e1.

It tries to specify latest version with the `v` tag with the longest
specificity.  actions/cache@v4 becomes actions/cache@v4.0.2, instead of
detecting that ref v4 also points to v4.0.2.  It would be nice to maintain the
same specificity of the tag semver as long as it points to the same git ref.
Otherwise one needs to run this update actions version workflow frequently to
mitigate increased specificity.  No, won't do; rather continue to update by
hand.

https://github.com/orgs/community/discussions/48058